code/docker/pip3.5-requirement.txt
PyYAML==3.12
simplejson==3.8.2
cython==0.25.2
pandas==0.19.0
sympy==1.0
futures==3.0.5
line-bot-sdk
flask==0.12
